Removed comments and console.log()
