+++ app/src/Homepage/Homepage-components.tsx
@@ -39,10 +39,7 @@ export const Header: React.FC<ChildrenProps> = ( props: ChildrenProps ) => {
 
 export const Logo: React.FC = () => {   
     
-    const handleClick = () => {
-        console.log('Button clicked: ' + this);
-        
-    };
+    const handleClick = () => {};
     
     return (
         <h1 className="logo" onClick={ handleClick }>
