+++ app/src/Homepage/Homepage.css
@@ -10,7 +10,6 @@
   position: fixed;
   z-index: 1;
   display: flex;
-  /* flex-direction: column; */
   justify-content: flex-start;
   align-items: center;
 }
