+++ reports/report_3.md
@@ -1,0 +1,81 @@
+# Report 3 (Week 6) due 11/01/23
+
+## Team Report
+
+### Previous Week Goals
+
+- Finish course page design
+- Complete milestone planned for 10/27/23
+    - Finish implementing the home page
+    - Finish implementing the directory pages
+- Complete Git, Testing, and CI Assignment due 10/31/23
+    - Set up an initial CI system in GitHub Actions
+- Start on milestone originally planned for 10/31/23
+    - Start to set up the database
+    - Start to implement course pages
+
+### Progress and Issues
+
+Progress:
+- Finished setting up the CI using GitHub Actions and the CI documentation
+- Course page design (and all others) completed
+- Set up routing for each page
+- The database is set up and populated with the data for courses
+- Homepage close to being done (all tasks assigned and being worked on)
+
+Issues:
+- We are a little behind as we are not done with the home page. We are close to finishing the home page but Thao is sick so we are down a developer.
+- The directory pages are not yet implemented. We are now working on figuring out how to get the data from our database to our application.
+- Since we are waiting on the directory to be implemented, implementing course pages will have to wait
+- We discovered that GitHub pages won’t be able to host our website given that we are storing data on AWS, working on finding new host
+
+### Next Week Goals
+
+- Get Beta Release ready
+    - Initial email login and verification set up
+    - Course directory implemented 
+        - Pages fetching course data from database
+        - Links to course pages
+    - Course pages done
+        - Pages fetching course data from database
+    - People can add a review
+        - Review UI
+        - Submitted reviews added to database
+
+## Individual Contributions
+
+### Previous Week Goals
+
+**Oliver:** Continue working on UI implementation, especially the login page.
+
+**Henry:** Finish design for Classes page, figure out the React issue.
+
+**Thao:** Continue working on UI implementation, maybe start putting in UW CSE courses into the database.
+
+**Celestine:** Continue working on UI implementation, especially the homepage. Hope to get started on programming directory pages and the API for database requests once the homepage is complete.
+
+**Jeremy:** Start to get the database set up on AWS. Main goal is to get Henry’s environment properly set up since he is the last one that is not able to.
+
+### Progress and Issues
+
+**Oliver:** Worked on implementing the Homepage and the Login page.
+
+**Henry:** Worked on cleaning up the classes data, created ratings page design and updated colors on figma using variables.
+
+**Thao:** Homepage is nearly fully implemented.
+
+**Celestine:** I have implemented page routing for our app and set up the sidebar buttons routing to directory pages. There were some issues as tests needed to be added and changed because of these big changes. I got the CI milestone finished and submitted. I also reviewed some PRs.
+
+**Jeremy:** I got the CI and database set up for us. I have also been on the lookout for PRs and getting them done as soon as possible. I was also able to fix the npm install and start issues we had in our repo due to a bad merge.
+
+### Next Week Goals
+
+**Oliver:** Have the sign in and email verification set up for the demo
+
+**Henry:**  Sort out UI issues in the design
+
+**Thao:** Have the homepage ready to demo
+
+**Celestine:** Have the class directory and course pages ready to demo
+
+**Jeremy:** Set up the database interaction for course directory, pages, and login set up. Support Celestine in getting the class directory and course pages ready to demo.
